Refactor goal setting functionality into a mixin
